ft_printf

+++ ft_conv_unsigned_int.c
@@ -5,8 +5,8 @@
 /*                                                    +:+ +:+         +:+     */
 /*   By: orazafin <orazafin@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2017/05/05 11:51:07 by orazafin          #+#    #+#             */
-/*   Updated: 2017/05/11 15:31:46 by orazafin         ###   ########.fr       */
+/*   Created: 2017/05/11 16:48:24 by orazafin          #+#    #+#             */
+/*   Updated: 2017/05/11 16:49:17 by orazafin         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -92,7 +92,7 @@ static int		display_padding_and_precision(unsigned int nb, t_option
 	return (result);
 }
 
-int				ft_conv_unsigned_int(t_option *option, long long nb)
+int				ft_conv_unsigned_int(t_option *option, unsigned int nb)
 {
 	unsigned int	decimal;
 	int				result;
